wifi router update/requirements.txt
flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
